The evmChainMapping variable in components/Activejs is derived from an old naming convention and has now been replaced by families.exe in config/mappingRegistry.js.
The evmChainMapping variable in components/Activejs is derived from an old naming convention and has now been replaced by families.exe in config/mappingRegistry.js. We have replaced the use of evmChainMapping in components/Activity. js with families. evm and maintained the correct reference to assetOps. After this correction, the chain identifier mapping and path mapping in activity.js are correctly matched, solving the problem caused by updating the names and methods of related files in the config folder but not synchronously updating activity.js.