The confirmDeleted Wallet function has been modified in components/SecureDevice. js to set the saved and persistent verifiedDevices to an empty array [] when calling this function.
The confirmDeleted Wallet function has been modified in components/SecureDevice. js to set the saved and persistent verifiedDevices to an empty array [] when calling this function. The specific implementation is to first call setVerifiedDevices ([]) to clear the state, then call Asynchronous Storage. setItem to persist the empty array, and then perform the wallet deletion operation and close the deletion confirmation mode box.