the updated cryptoCards will be used to ensure that prices and balances are refreshed correctly.
The reason why fetchPriceChanges and fetchWalletBalance are not executed in the handleAddCrypto function is that the cryptoCards parameter passed in when they are called is the old state value. The parameters of these two function calls have been changed to newCryptoCards, which is the latest state value. In this way, when calling fetchPriceChanges and fetchWalletBalance, the updated cryptoCards will be used to ensure that prices and balances are refreshed correctly.